[front/lib/api] - refactor: remove handlePatchDataSourceView function
 - Relocated handlePatchDataSourceView from data_source_view.ts to its own API endpoint module

[front/pages/api] - refactor: relocate handlePatchDataSourceView to specific API module

 - Transferred handlePatchDataSourceView function for better modularity and encapsulation within the API directory structure
