+++ front/pages/api/v1/w/[wId]/vaults/[vId]/data_source_views/[dsvId]/index.ts
@@ -1,11 +1,11 @@
 import type { DataSourceViewType, WithAPIErrorResponse } from "@dust-tt/types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { handlePatchDataSourceView } from "@app/lib/api/data_source_view";
 import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { apiError } from "@app/logger/withlogging";
+import { handlePatchDataSourceView } from "@app/pages/api/w/[wId]/vaults/[vId]/data_source_views/[dsvId]";
 
 export type GetOrPostDataSourceViewsResponseBody = {
   dataSourceView: DataSourceViewType;
